update to mcptest, add Distance Extender
distance extender as suggested by stellardawn

I originally had ChunkProviderWDL inserted in between WorldClient and
ChunProviderClient, which kept track of its own chunkmap/list, but
decided it was best to just have it all in one class.  Could have edited
ChunkProviderClient, but I decided not to edit one more base class, so a
lot of ChunkProviderClient is replicated in ChunkProviderWDL which may
not be any better.  It is a pretty simple class at any rate.  And the
whole thing could likely be done better, but it works (SP too!)