liteloader
add files to create litemod version of World Downloader.  Together with
the files in /wdl, they make a complete mod (no base class edits
needed).  That can be either WDL as it currently is, or with the
voxelmap/observer changes I suggested earlier.  The litemod stuff is
purely a drop in replacement for all the previous base class edits